requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Financial data and analysis
yfinance>=0.1.70
alpha-vantage>=2.3.1
ta>=0.10.1

# Real-time data and websockets
websocket-client>=1.2.0
aiohttp>=3.8.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0

# Logging and monitoring
loguru>=0.6.0

# Testing
pytest>=6.2.0
pytest-asyncio>=0.18.0

# Code quality
black>=22.0.0
flake8>=4.0.0

# Time handling
python-dateutil>=2.8.0
pytz>=2021.3

# HTTP requests
requests>=2.26.0
urllib3>=1.26.0

# Async support
aiofiles>=0.8.0

# Statistical analysis
statsmodels>=0.13.0 